Ai voice (#2)
* local.properties file being added to git repo

* fixing hint text

* fixing ai voice to extract weight & height from anywhere in the statement

* adding unit test case

* writing modular code

* writing modular code

* writing modular code

* writing modular code

* fixing defect in height

* fixing defect in height

* fixing defect in height

* fixing defect in height

+++ app/src/main/java/com/ritesh/bmi/calculator/output/EBMICategory.java
@@ -14,21 +14,21 @@ import java.util.Arrays;
  * 39.9 < Obese Class 3 / Severe Obesity
  */
 public enum EBMICategory {
-    UNKNOWN("Unknown", -1, -1, Color.BLACK),
-    UNDERWEIGHT("Underweight", 0, 18.5, Color.BLUE),
-    NORMAL_WEIGHT("Normal weight", 18.5, 24.9, Color.GREEN),
-    OVERWEIGHT("Overweight", 24.9, 29.9, Color.rgb(250, 197, 5)),
-    OBESE_1("Obese Class 1", 29.9, 34.9, Color.rgb(255, 165, 0)),  //Orange
-    OBESE_2("Obese Class 2", 34.9, 39.9, Color.RED),
-    OBESE_3("Obese Class 3 / Severe Obesity", 39.9, Double.MAX_VALUE, Color.rgb(128, 0, 128)); //Purple
+    UNKNOWN("Unknown", -1f, -1f, Color.BLACK),
+    UNDERWEIGHT("Underweight", 0f, 18.5f, Color.BLUE),
+    NORMAL_WEIGHT("Normal weight", 18.5f, 24.9f, Color.GREEN),
+    OVERWEIGHT("Overweight", 24.9f, 29.9f, Color.rgb(250, 197, 5)),
+    OBESE_1("Obese Class 1", 29.9f, 34.9f, Color.rgb(255, 165, 0)),  //Orange
+    OBESE_2("Obese Class 2", 34.9f, 39.9f, Color.RED),
+    OBESE_3("Obese Class 3 / Severe Obesity", 39.9f, Float.MAX_VALUE, Color.rgb(128, 0, 128)); //Purple
 
     private final String displayName;
-    private final double lowerValue;
-    private final double upperValue;
+    private final float lowerValue;
+    private final float upperValue;
 
     private final int color;
 
-    EBMICategory(String displayName, double lowerValue, double upperValue, int color) {
+    EBMICategory(String displayName, float lowerValue, float upperValue, int color) {
         this.displayName = displayName;
         this.lowerValue = lowerValue;
         this.upperValue = upperValue;
